worker/requirements.txt
amqp==5.1.1
anyio==3.3.4
asgiref==3.4.1
async-timeout==4.0.2
billiard==3.6.4.0
cachetools==4.2.4
celery==5.2.7
celerybeat-mongo==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dnspython==2.1.0
fastapi==0.70.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
h11==0.12.0
httplib2==0.20.2
idna==3.3
joblib==1.1.0
kombu==5.2.4
mongoengine==0.24.1
nltk==3.7
packaging==21.3
prompt-toolkit==3.0.29
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pymongo==3.12.1
pyparsing==3.0.6
python-dotenv==0.20.0
pytz==2022.1
rabbitmq==0.2.0
rake-nltk==1.0.6
redis==4.3.3
regex==2022.6.2
requests==2.26.0
rsa==4.7.2
schedule==1.1.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tqdm==4.64.0
typing-extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
